Eight commit

+++ Python for Data Professional Beginner/Part 3/20.tugas-praktek-4.py
@@ -18,7 +18,7 @@ class TenagaLepas(Karyawan):
     def __init__(self, nama, usia, pendapatan):
         super().__init__(nama, usia, pendapatan, 0)
     def tambahan_proyek(self, nilai_proyek):
-        self.pendapatan_tambahan += nilai_proyek * 0.01
+        self.pendapatan_tambahan += int(nilai_proyek * 0.01)
 # Definisikan class AnalisData sebagai child class dari
 # class Karyawan
 class AnalisData(Karyawan): 
@@ -27,7 +27,7 @@ class AnalisData(Karyawan):
 # class Karyawan
 class IlmuwanData(Karyawan):
     def tambahan_proyek(self, nilai_proyek): 
-        self.pendapatan_tambahan += 0.1 * nilai_proyek
+        self.pendapatan_tambahan += int(0.1 * nilai_proyek)
 # Definisikan class PembersihData sebagai child class dari
 # class TenagaLepas
 class PembersihData(TenagaLepas):
